refactor(core): adjust float accuracy to f32; feat(core): use glam to replace Point implement

+++ crates/core/src/traits/focusable.rs
@@ -3,5 +3,5 @@ use super::RendererUpdatePayload;
 pub trait Focusable {
     /// check whether the input point is on the node-like,
     /// coordinate of the point is relative to its parent
-    fn contains(&self, x: f64, y: f64, payload: &RendererUpdatePayload) -> bool;
+    fn contains(&self, x: f32, y: f32, payload: &RendererUpdatePayload) -> bool;
 }
